fixes 1577 and zymo JSON w/r

+++ protocols/zymo-ribofree-p5-adapter-ligation/3_p5_adapter_ligation.ot2.apiv2.py
@@ -43,19 +43,20 @@ def run(ctx):
     p20.flow_rate.blow_out = 30
     m50 = ctx.load_instrument('p50_multi', p50_mount, tip_racks=racks50)
 
-    file_path = '/data/csv/tip_track.json'
-    # file_path = 'protocols/tip_track.json'
-    if os.path.isfile(file_path):
-        with open(file_path) as json_file:
-            data = json.load(json_file)
-            if 'tips20' in data:
-                tip20_count = data['tips20']
-            else:
-                tip20_count = 0
-            if 'tips50' in data:
-                tip50_count = data['tips50']
-            else:
-                tip50_count = 0
+    if not ctx.is_simulating():
+        file_path = '/data/csv/tip_track.json'
+        # file_path = 'protocols/tip_track.json'
+        if os.path.isfile(file_path):
+            with open(file_path) as json_file:
+                data = json.load(json_file)
+                if 'tips20' in data:
+                    tip20_count = data['tips20']
+                else:
+                    tip20_count = 0
+                if 'tips50' in data:
+                    tip50_count = data['tips50']
+                else:
+                    tip50_count = 0
     else:
         tip20_count = 0
         tip50_count = 0
@@ -126,9 +127,10 @@ with cleanup.')
 
     # track final used tip
     # file_path = '/data/csv/tip_track.json'
-    data = {
-        'tips20': tip20_count,
-        'tips50': tip50_count
-    }
-    with open(file_path, 'w') as outfile:
-        json.dump(data, outfile)
+    if not ctx.is_simulating():
+        data = {
+            'tips20': tip20_count,
+            'tips50': tip50_count
+        }
+        with open(file_path, 'w') as outfile:
+            json.dump(data, outfile)
